Various fixes and implementations
Buffer for points that are to be drawn
Moved the drawing code to the widget draw function, drawing from the buffer
The preview window is now sharing the context with the image area but there are some problems
Use f64 instead of f32 for accuracy and skipping some casts, though introducing others
Fixed an undetected problem in the local code of the input system

+++ src/state.rs
@@ -12,8 +12,8 @@ pub struct State {
     pub palettes: Vec<Color>,//TODO: Multiple palettes
     pub input: Vec<Input>,
     pub args: Vec<Arg>,
-    pub zoom: f32,
-    pub translation: [f32; 2],
+    pub zoom: f64,
+    pub translation: [f64; 2],
 	pub input_buffer: String,
 }
